Add login and logout js files, moved and updated handlebars
