+++ routes/index.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 
-const apiRoutes = require('./api');
+const apiRoutes = require('../controllers/api');
 
 router.use('/api', apiRoutes);
