--- views/all.handlebars
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" class= "bg-yellow-100	">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <script src="https://cdn.tailwindcss.com"></script>
-    <script>
-        tailwind.config = {
-          theme: {
-            extend: {
-              colors: {
-                clifford: '#da373d',
-              }
-            }
-          }
-        }
-      </script>
-    <title>CBDB</title>
-</head>
-<body>
-    <header class="grid flex justify-start">CBDB</header>
-</body>
-</html>
